# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "DEVinBank Pagamentos S.A - API Conta365",
    "description": "API para gerenciar gastos pessoais",
    "version": "1.0.0"
  },
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/financial/{userID}": {
      "get": {
        "tags": [
          "Financial"
        ],
        "description": "Endpoint que lista as despesas por ano/mês de um usuário específico (ID)",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "typesOfExpenses",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Financial"
        ],
        "description": "Endpoint que adiciona despesas na lista de gastos de um determinado usuário (ID)",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Some description...",
            "accept": "/"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/financial/{userID}/{financialID}": {
      "delete": {
        "tags": [
          "Financial"
        ],
        "description": "Endpoint que remove de um determinado usuário (ID) uma despesa específica (ID)",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "financialID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}